requirements.txt
apscheduler==3.11.0
blinker==1.9.0
brotli==1.1.0
cachelib==0.13.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
crontab==1.0.4
cssselect2==0.8.0
cycler==0.12.1
flask==3.1.0
flask-caching==2.3.1
flask-cors==5.0.1
fonttools==4.57.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.8
markdown==3.8
markupsafe==3.0.2
matplotlib==3.10.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyparsing==3.2.3
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
werkzeug==3.1.3
zopfli==0.2.3.post1

# Azure Dependencies
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.19.0
opencensus-ext-azure==1.1.13
opencensus-ext-flask==0.8.3
opencensus-ext-logging==0.1.1
